Addressed nits, moved delete operator into a separate file.
